refactor: break up monolithic routes.py into modular domain structure
