+++ application/api/user/analytics/__init__.py
@@ -1,0 +1,5 @@
+"""Analytics module."""
+
+from .routes import analytics_ns
+
+__all__ = ["analytics_ns"]
