+++ application/api/user/attachments/__init__.py
@@ -1,0 +1,5 @@
+"""Attachments module."""
+
+from .routes import attachments_ns
+
+__all__ = ["attachments_ns"]
