+++ application/api/user/conversations/__init__.py
@@ -1,0 +1,5 @@
+"""Conversation management module."""
+
+from .routes import conversations_ns
+
+__all__ = ["conversations_ns"]
