+++ application/api/user/prompts/__init__.py
@@ -1,0 +1,5 @@
+"""Prompts module."""
+
+from .routes import prompts_ns
+
+__all__ = ["prompts_ns"]
