+++ application/api/user/sharing/__init__.py
@@ -1,0 +1,5 @@
+"""Sharing module."""
+
+from .routes import sharing_ns
+
+__all__ = ["sharing_ns"]
